+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 7.20.0
+
+### Minor Changes
+
+- add moon-select package
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@7.20.0
+
 ## 7.19.1
 
 ### Patch Changes